+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0] 2023-06-10
+- Removed --tui-fold-level option, now replaced by modified --tui-regexfile implementation (previous bullet point); allows for regex(es) to be specified in a file, rather than on the command line.
+- Fixed Issue 100 (obviated by previous change in fold behavior).
+- Changed output files folder to ./tui_files.
+- Added new buttons to About tab to remove/restore/invert colors.
+- Introduced tests to test the plugin itself, using pytest's 'pytester' fixture.
+- New buttons in HTML report to invert/remove/restore colors.
+
 ## [1.10.0] 2023-04-12
 - Added new folding mode `--tui-fold-regex`. This mode allows specifying reg-ex's marking the beginning and end of a section that the user wants to fold. See README for details.
 - Bumped `rich` library version to 12.6.0 to allow for use of SeleniumBase library during testing.
@@ -42,7 +50,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - Tweaked and formatted a bunch of the tests in /demo-tests.
 ## [1.6.1] 2022-09-23
 
-- Added '*' .gitignore to ptt_files/ so that when people run pytest --tui in other directories they don't see the ptt_files/ dir.
+- Added '*' .gitignore to tui_files/ so that when people run pytest --tui in other directories they don't see the tui_files/ dir.
 
 ## [1.6.0] - 2022-09-20
 
@@ -113,7 +121,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - Fixed tui1 (Textual) so that it works with new internal implementation.
 - Retiring tui2 (PyTermTk) for now.
 - Reordered/removed some menu items in tuiconf to fit new content/choices.
-- Changed output files folder to ./ptt_files.
+- Changed output files folder to ./tui_files.
 
 ## [1.1.3] - 2022-08-13